remove extra config, move config to pkg/
