--- plugin/storage/clickhouse/options.go
@@ -1,4 +1,0 @@
-// Copyright (c) 2025 The Jaeger Authors.
-// SPDX-License-Identifier: Apache-2.0
-
-package clickhouse
